+++ zcash_transparent/src/builder.rs
@@ -8,7 +8,7 @@ use zcash_protocol::value::{BalanceError, ZatBalance, Zatoshis};
 
 use crate::{
     address::{Script, TransparentAddress},
-    bundle::{Authorization, Authorized, Bundle, TxIn, TxOut},
+    bundle::{Authorization, Authorized, Bundle, MapAuth, TxIn, TxOut},
     pczt,
     sighash::{SignableInput, TransparentAuthorizingContext},
 };
@@ -183,6 +183,19 @@ impl TransparentAuthorizingContext for Coinbase {
 
     fn input_scriptpubkeys(&self) -> Vec<Script> {
         vec![]
+    }
+}
+
+impl MapAuth<Coinbase, Authorized> for Coinbase {
+    fn map_script_sig(
+        &self,
+        s: <Coinbase as Authorization>::ScriptSig,
+    ) -> <Authorized as Authorization>::ScriptSig {
+        Script(s)
+    }
+
+    fn map_authorization(&self, _: Coinbase) -> Authorized {
+        Authorized
     }
 }